Restore a line of code I thought was unnecessary
Experiment concluded. `Skip(1)` is needed in the case that a .WAD file is passed in directly first, and `TakeWhile()` is needed in other cases. I think that's basically correct.

It failed on me for the first time when using `-file` to pass in the .WAD I wanted to play; it was the first argument and the other arguments caused an exception.