Comment `DoEvents()`
This one appears pretty simple. It goes through the event queue and coordinates other parts of the engine to handle the events meant for them. There is a menu object for menu events, a place to handle function keys, etc.

Like it says in GEBB:DOOM, events are passed around to different systems and consumed (or "eaten") if a given system handles that event.