Comment `Update()`
Surprisingly simple. It runs `DoEvents()` then handles screen wiping. Then if the screen is not wiping, it handles state transitions and other objects' game loops. At the end of a full execution, it runs `CheckMouseState` and returns `UpdateResult.None`, which represents a normal frame (as opposed to a frame where the game was quit).