Comment `ResumeGame()`
What stands out to me here is that `sendPause` is set to true whether you are pausing or unpausing the game. I assume that means the flag is simply checked to see if the pause state should be toggled each frame before resetting the flag. I like that. It simplifies an idea I had for event-based game design on a recent project I was working on.